[Feat] NavigationBar appearance를 전역적으로 적용
