+++ Favor/Favor/Sources/Scenes/Application/SceneDelegate.swift
@@ -14,7 +14,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
 	func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
 		guard let windowScene = (scene as? UIWindowScene) else { return }
-		let navigationController = BaseNavigationController()
+		let navigationController = UINavigationController()
 		
 		self.window = UIWindow(windowScene: windowScene)
 		self.window?.rootViewController = navigationController
